# examples/typed-client/openapi.yaml
openapi: 3.0.3
info:
  title: Items API (okapi example)
  version: 0.1.0
  description: Tiny fixture used by the client-test example.

paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /items:
    get:
      operationId: listItems
      summary: List items
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 200
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemList"
    post:
      operationId: createItem
      summary: Create an item
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /items/export:
    get:
      operationId: exportItems
      summary: Export items as CSV
      responses:
        "200":
          description: CSV stream
          content:
            text/csv: {}

  /items/import:
    post:
      operationId: importItems
      summary: Bulk-import items from a CSV upload
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: No Content

  /items/{id}:
    get:
      operationId: getItem
      summary: Get an item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteItem
      summary: Delete an item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
        - in: header
          name: If-Match
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content

components:
  schemas:
    Health:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
      required: [status]

    Item:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        created_at:
          type: string
      required: [id, name, created_at]

    ItemList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
      required: [items]

    CreateItemBody:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
      required: [name]

    Problem:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required: [title, status]